s2257, ta9390: rename SmokaResolver to SubaruResolver, + rename related tests.
