+++ caom2-artifact-resolvers/src/test/java/ca/nrc/cadc/caom2/artifact/resolvers/SubaruResolverTest.java
@@ -69,14 +69,9 @@
 
 package ca.nrc.cadc.caom2.artifact.resolvers;
 
-import ca.nrc.cadc.net.HttpDownload;
 import ca.nrc.cadc.util.Log4jInit;
-import java.io.ByteArrayOutputStream;
-import java.io.OutputStream;
 import java.net.URI;
 import java.net.URL;
-import java.util.ArrayList;
-import java.util.List;
 import org.apache.log4j.Level;
 import org.apache.log4j.Logger;
 import org.junit.Assert;
@@ -85,45 +80,114 @@ import org.junit.Test;
 /**
  * @author hjeeves
  */
-public class SmokaResolverIntTest {
-    private static final Logger log = Logger.getLogger(SmokaResolverIntTest.class);
+public class SubaruResolverTest {
+    private static final Logger log = Logger.getLogger(SubaruResolverTest.class);
 
     static {
         Log4jInit.setLevel("ca.nrc.cadc", Level.INFO);
     }
 
-    String VALID_DATA_URI = "subaru:data/2017-01-01/SUPE01318470";
-    String VALID_PREVIEW_URI = "subaru:preview/SUPE01318470";
+    String VALID_FILE1 = "SUPE01318470";
+    String VALID_DATE1 = "2017-09-09";
+    String VALID_FILE2 = "SUPE01318470";
 
-    SmokaResolver resolver = new SmokaResolver();
+    String SCHEME = "subaru";
+    String DATA_URI = "data";
+    String PREVIEW_URI = "preview";
 
-    public SmokaResolverIntTest() {
+    String PROTOCOL_STR = "http://";
+
+    String BASE_PREVIEW_URL = "smoka.nao.ac.jp";
+    String PREVIEW_URL_QUERY = "grayscale=linear&mosaic=true&frameid=";
+    String PREVIEW_URL_PATH = "/qlis/ImagePNG";
+
+    String BASE_DATA_URL = "www.canfar.net";
+    String DATA_URL_PATH = "/maq/subaru";
+    String DATA_URL_QUERY= "frameinfo=";
+
+    // Invalid checks the scheme and the request type (needs to be 'file' or 'preview')
+    String INVALID_URI_BAD_SCHEME = "pokey:little/puppy";
+
+    SubaruResolver subaruResolver = new SubaruResolver();
+
+    public SubaruResolverTest() {
     }
 
     @Test
-    public void testValidSiteUrl() throws Exception {
-        log.info("starting testValidSiteUrl");
+    public void testGetSchema() {
+        Assert.assertTrue(subaruResolver.getScheme().equals(subaruResolver.getScheme()));
+    }
 
+    @Test
+    public void testValidURI() {
         try {
-            List<URI> uriList = new ArrayList<URI>();
-            uriList.add(new URI(VALID_DATA_URI));
-            uriList.add(new URI(VALID_PREVIEW_URI));
+            String uriStr = subaruResolver.getScheme() + ":" + DATA_URI + "/" + VALID_DATE1 + "/" + VALID_FILE1;
+            URI uri = new URI(uriStr);
+            URL url = subaruResolver.toURL(uri);
+            log.debug("toURL returned: " + url.toString());
 
-            for (URI uri : uriList) {
-                URL url = resolver.toURL(uri);
+            Assert.assertEquals(url.toString(), PROTOCOL_STR + BASE_DATA_URL + DATA_URL_PATH + "?" + DATA_URL_QUERY + VALID_FILE1 + "%20" + VALID_DATE1);
+            Assert.assertEquals(DATA_URL_PATH, url.getPath());
+            Assert.assertEquals(DATA_URL_QUERY + VALID_FILE1 + "%20" + VALID_DATE1, url.getQuery());
+            Assert.assertEquals(BASE_DATA_URL, url.getHost());
 
-                log.debug("opening connection to: " + url.toString());
-                OutputStream out = new ByteArrayOutputStream();
-                HttpDownload head = new HttpDownload(url, out);
-                head.setHeadOnly(true);
-                head.run();
-                Assert.assertEquals(200, head.getResponseCode());
-                log.info("response code: " + head.getResponseCode());
-            }
+            uriStr = subaruResolver.getScheme() + ":" + PREVIEW_URI + "/" + VALID_FILE2;
+            uri = new URI(uriStr);
+            url = subaruResolver.toURL(uri);
+            log.debug("toURL returned: " + url.toString());
+            // http://smoka.nao.ac.jp/qlis/ImagePNG?grayscale=linear&mosaic=true&frameid=SUPE01318470
+
+            Assert.assertEquals(url.toString(), PROTOCOL_STR + BASE_PREVIEW_URL + PREVIEW_URL_PATH + "?" + PREVIEW_URL_QUERY + VALID_FILE2);
+            Assert.assertEquals(PREVIEW_URL_PATH, url.getPath());
+            Assert.assertEquals(PREVIEW_URL_QUERY + VALID_FILE2, url.getQuery());
+            Assert.assertEquals(BASE_PREVIEW_URL, url.getHost());
 
         } catch (Exception unexpected) {
-            log.error("Unexpected exception", unexpected);
-            Assert.fail("Unexpected exception: " + unexpected);
+            log.error("unexpected exception", unexpected);
+            Assert.fail("unexpected exception: " + unexpected);
         }
     }
+
+    @Test
+    public void testInvalidURIBadScheme() {
+        try {
+            URI uri = new URI(INVALID_URI_BAD_SCHEME);
+            URL url = subaruResolver.toURL(uri);
+            Assert.fail("expected IllegalArgumentException, got " + url);
+        } catch (IllegalArgumentException expected) {
+            log.info("IllegalArgumentException thrown as expected. Test passed.: " + expected);
+        } catch (Exception unexpected) {
+            log.error("unexpected exception", unexpected);
+            Assert.fail("unexpected exception: " + unexpected);
+        }
+    }
+
+    @Test
+    public void testInvalidNullURI() {
+        try {
+            URL url = subaruResolver.toURL(null);
+            Assert.fail("expected IllegalArgumentException, got " + url);
+        } catch (IllegalArgumentException expected) {
+            log.info("IllegalArgumentException thrown as expected. Test passed.: " + expected);
+        } catch (Exception unexpected) {
+            log.error("unexpected exception", unexpected);
+            Assert.fail("unexpected exception: " + unexpected);
+        }
+    }
+
+    @Test
+    public void testInvalidUriType() {
+        try {
+            String uriStr = GeminiResolver.SCHEME + ":badURIType/" + VALID_FILE1;
+            URI uri = new URI(uriStr);
+            URL url = subaruResolver.toURL(uri);
+            Assert.fail("expected IllegalArgumentException, got " + url);
+        } catch (IllegalArgumentException expected) {
+            log.info("IllegalArgumentException thrown as expected. Test passed.: " + expected);
+        } catch (Exception unexpected) {
+            log.error("unexpected exception", unexpected);
+            Assert.fail("unexpected exception: " + unexpected);
+        }
+    }
+
 }
